replace with correct script from Feb 4 version
Looks like I accidentally replaced this script `similarities_in_proteinprotein_interactions.py` with my April 9th updated version of `get_seq_from_multiFASTA_with_match_in_description.py`.  
This rolls back to the prior February 4th version of this script because the present version I have in my active Dropbox (and locally) has no differences from that version. I ran `diff` to see if this was the case and it was because I didn't know that before I did that.