add draft of code to switch from use of curl to requests for getting code files at GitHub
Not using it here yet but noting it since it is one place I should update if seems to work well. That would make it more robust as it would work in more places easily as I'm not sure about `from sh import curl` on Windows 10. And I know previous Windows versions don't have curl included.